+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changelog
 - Add ``over``. Thanks bharadwajyarlagadda_!
 - Add ``over_every``. Thanks bharadwajyarlagadda_!
 - Add ``over_some``. Thanks bharadwajyarlagadda_!
+- Add ``pull_all``. Thanks bharadwajyarlagadda_!
 - Add ``range_right``. Thanks bharadwajyarlagadda_!
 - Add ``sample_size``. Thanks bharadwajyarlagadda_!
 - Add ``sorted_uniq``. Thanks bharadwajyarlagadda_!